+++ parquetx/schema_test.go
@@ -45,6 +45,15 @@ func TestSchemaFromMessageDescriptor(t *testing.T) {
                 }
 			`),
 		},
+		{
+			"enum field",
+			(&pbtesting.RowColumEnum{}).ProtoReflect().Descriptor(),
+			schemaLiteral(`
+                message rows {
+                  required binary value (ENUM);
+                }
+			`),
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {